requirements.txt
Flask
mysql-connector-python
flask-jwt-extended
bcrypt
flasgger
